jobify_with_PuchAI/requirements.txt
fastmcp
pdfminer.six
markdownify
readabilipy
httpx
pydantic
python-dotenv